Update build script to add linker runpath into executables
This avoids any needs to set LD_LIBRARY_PATH or editing of ld.so.conf files.
The runpath is set to ../lib, so dynamic library must reside there relative 
to the executables for this to work. 

Tested on Ubuntu 12.04, needs testing on MacOSX. 
No need for testing on Windows, since dll search path includes folder where 
executable resides.